issue #7 and #8 Hopefully put to bed for good now. issue #14 also fixed
issue #7 A couple of issues but seems to be fixed now. Fixed the DISPLAYDLL_CustomCharIndex dll function handler
issue #8 Is now fixed. Thanks to limbo666 and rdoursenaud for testing and confirming this.
Issue #14 A bit of silliness in the way the client area was calculated. fix by manually specifying paintbox size
Also, added project for a new display dll. It's part of an android screen app I'm developing but at this time its insecure so it is not being built and distributed in the release zip yet.