A few fixes
Setup form has minimum size.
Setup form size saved in config.
Add button for backlight in setup form.
It was impossible to browse list boxes with arrow keys so removed a function that focuses the input field when selecting an item from list boxes. It needs testing to make sure the line and position input to is correct.
Added new plugin functions SmartieInfo() and SmartieDemo() for fetching info and demos for display in the setup form.
Updated DNBridge to handle new plugin functions
Updated LegacyLoader to handle new plugin functions