Some suggestions/fixes
Some suggestions/fixes reported by limbo
If the dll is not exposing the info (modern) and if the info file is missing the program should be able add a basic line
The plugins list Refresh button need to be repositioned
The list if action selection should be updated. EnableScreen(1-20) should be updated to the maximum screen number that LCD Smartie supports
The Actions list is missing a Duplicate button
If you check the "Stay on top" checkbox and click Apply on the configuration window everything goes fine. If you uncheck it and click Apply then the configuration window is closes
Config window title should be updated to display current program version